Extract controls js_only
